Update overall progress
A lot of pieces are moving internally, but updating the document is quite hard due to it all being in a single document. At some point I hope to split this out into individual issues, but that may come for future work, rather than the existing work. Right now we're focused on getting-things-done (and things are definitely getting done behind the scenes).