Document multiplayer score submission flow
This PR adds documentation about the multiplayer score submission flow
in the form of a sequence diagram. It was primarily created by taking
the existing solo score submission flow sequence diagram and adapting
it to the differences specific to the multi flow (different endpoints,
different tables, different ID schemes).

This diagram does not describe the current state of things. It describes
where things _should_ be after a sequence of some already-PRed changes
such as:

- https://github.com/ppy/osu-web/pull/10437
- https://github.com/ppy/osu/pull/24697
- https://github.com/ppy/osu-server-spectator/pull/185

and some changes that are still being worked on (primarily client-side
at this point).

The main goal of the document is to inform the changes above. (And to be
something that I can easily refer to, because I keep looking up the same
information over and over right now.)